attempt to make docker-gen report all containers even if stopped
stopping a zookeeper container should not cause the other running containers
to be re-generated and restarted. however, at the moment that is exactly what
happens.. when stopping a container, docker-gen triggers and does not pass the
stopped container in the emitted data set so the templates rebuild the server list
without the stopped container present. then docker-gen triggers the ZK restart..
net result is stopping a container, removes the container from the cluster

ofcourse starting the container brings it back into the container, so it's not all bad
but ideally if docker-gen passed stopped containers it would recognize that the
config files haven't changed and therefore would not restart the running zk instances
which is ultmately a better option.

longer term the dynamic config settings in the zk 3.5 branch  should be used instead of this